KREST-3637 Address first round of review comments
- return get schema registry client interface to how it was
- get rid of the multi exception
- always throw an error if there is no schema regisry client but a schema is specified (except if the format is json or binary, and that is checked for in produceaction)
