+++ kafka-rest/src/main/java/io/confluent/kafkarest/DefaultKafkaRestContext.java
@@ -24,7 +24,6 @@ import io.confluent.kafkarest.config.SchemaRegistryConfig;
 import io.confluent.kafkarest.v2.KafkaConsumerManager;
 import java.net.URI;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.admin.AdminClient;
@@ -92,10 +91,10 @@ public class DefaultKafkaRestContext implements KafkaRestContext {
   }
 
   @Override
-  public Optional<SchemaRegistryClient> getSchemaRegistryClient() {
+  public SchemaRegistryClient getSchemaRegistryClient() {
 
     if (!config.isSchemaRegistryEnabled()) {
-      return Optional.empty();
+      return null;
     }
     if (schemaRegistryClient == null) {
       SchemaRegistryConfig schemaRegistryConfig =
@@ -114,7 +113,7 @@ public class DefaultKafkaRestContext implements KafkaRestContext {
               config.getSchemaRegistryConfigs(),
               schemaRegistryConfig.requestHeaders());
     }
-    return Optional.of(schemaRegistryClient);
+    return schemaRegistryClient;
   }
 
   @Override
