+++ kafka-rest/src/main/java/io/confluent/kafkarest/KafkaRestContext.java
@@ -17,7 +17,6 @@ package io.confluent.kafkarest;
 
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.confluent.kafkarest.v2.KafkaConsumerManager;
-import java.util.Optional;
 import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.producer.Producer;
 
@@ -37,7 +36,7 @@ public interface KafkaRestContext {
     return getProducerPool().getProducer();
   }
 
-  default Optional<SchemaRegistryClient> getSchemaRegistryClient() {
+  default SchemaRegistryClient getSchemaRegistryClient() {
     return null;
   }
