+++ kafka-rest/src/main/java/io/confluent/kafkarest/backends/schemaregistry/SchemaRegistryModule.java
@@ -19,6 +19,13 @@ import static java.util.Objects.requireNonNull;
 
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.confluent.kafkarest.KafkaRestContext;
+import io.confluent.kafkarest.config.ConfigModule.AvroSerializerConfigs;
+import io.confluent.kafkarest.config.ConfigModule.JsonschemaSerializerConfigs;
+import io.confluent.kafkarest.config.ConfigModule.ProtobufSerializerConfigs;
+import io.confluent.kafkarest.controllers.SchemaRecordSerializer;
+import io.confluent.kafkarest.controllers.SchemaRecordSerializerImpl;
+import io.confluent.kafkarest.controllers.SchemaRecordSerializerThrowing;
+import java.util.Map;
 import java.util.Optional;
 import javax.inject.Inject;
 import javax.inject.Provider;
@@ -38,6 +45,10 @@ public final class SchemaRegistryModule extends AbstractBinder {
     bindFactory(SchemaRegistryClientFactory.class)
         .to(new TypeLiteral<Optional<SchemaRegistryClient>>() {})
         .in(RequestScoped.class);
+
+    bindFactory(SchemaRecordSerializerFactory.class)
+        .to(SchemaRecordSerializer.class)
+        .in(RequestScoped.class);
   }
 
   private static final class SchemaRegistryClientFactory
@@ -52,10 +63,47 @@ public final class SchemaRegistryModule extends AbstractBinder {
 
     @Override
     public Optional<SchemaRegistryClient> provide() {
-      return context.get().getSchemaRegistryClient();
+      return Optional.ofNullable(context.get().getSchemaRegistryClient());
     }
 
     @Override
     public void dispose(Optional<SchemaRegistryClient> schemaRegistryClient) {}
+  }
+
+  private static final class SchemaRecordSerializerFactory
+      implements Factory<SchemaRecordSerializer> {
+
+    private final Optional<SchemaRegistryClient> schemaRegistryClient;
+    private final Map<String, Object> avroSerializerConfigs;
+    private final Map<String, Object> jsonschemaSerializerConfigs;
+    private final Map<String, Object> protobufSerializerConfigs;
+
+    @Inject
+    private SchemaRecordSerializerFactory(
+        Optional<SchemaRegistryClient> schemaRegistryClient,
+        @AvroSerializerConfigs Map<String, Object> avroSerializerConfigs,
+        @JsonschemaSerializerConfigs Map<String, Object> jsonschemaSerializerConfigs,
+        @ProtobufSerializerConfigs Map<String, Object> protobufSerializerConfigs) {
+      this.schemaRegistryClient = schemaRegistryClient;
+      this.avroSerializerConfigs = avroSerializerConfigs;
+      this.jsonschemaSerializerConfigs = jsonschemaSerializerConfigs;
+      this.protobufSerializerConfigs = protobufSerializerConfigs;
+    }
+
+    @Override
+    public SchemaRecordSerializer provide() {
+      if (schemaRegistryClient.isPresent()) {
+        return new SchemaRecordSerializerImpl(
+            schemaRegistryClient,
+            avroSerializerConfigs,
+            jsonschemaSerializerConfigs,
+            protobufSerializerConfigs);
+      } else {
+        return new SchemaRecordSerializerThrowing();
+      }
+    }
+
+    @Override
+    public void dispose(SchemaRecordSerializer schemaRecordSerializer) {}
   }
 }
